fix: convert to js
