--- frontend/Components/App.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-/** Components */
-import AppRouter from '../Navigation/Routes';
-
-/** Ant design stylesheet */
-import 'antd/dist/antd.css';
-
-const App = () => {
-  return <AppRouter />;
-};
-
-export default App;
-
