+++ frontend/amplify/#current-cloud-backend/amplify-meta.json
@@ -1,0 +1,56 @@
+{
+  "providers": {
+    "awscloudformation": {
+      "AuthRoleName": "amplify-myapp-dev-184113-authRole",
+      "UnauthRoleArn": "arn:aws:iam::427176174668:role/amplify-myapp-dev-184113-unauthRole",
+      "AuthRoleArn": "arn:aws:iam::427176174668:role/amplify-myapp-dev-184113-authRole",
+      "Region": "ap-southeast-2",
+      "DeploymentBucketName": "amplify-myapp-dev-184113-deployment",
+      "UnauthRoleName": "amplify-myapp-dev-184113-unauthRole",
+      "StackName": "amplify-myapp-dev-184113",
+      "StackId": "arn:aws:cloudformation:ap-southeast-2:427176174668:stack/amplify-myapp-dev-184113/e8d801a0-4ca1-11ee-b4d9-0abcacc0f598",
+      "AmplifyAppId": "d1pufa9ecg46gu"
+    }
+  },
+  "auth": {
+    "myapp1099f97a": {
+      "service": "Cognito",
+      "providerPlugin": "awscloudformation",
+      "dependsOn": [],
+      "customAuth": false,
+      "frontendAuthConfig": {
+        "socialProviders": [],
+        "usernameAttributes": [],
+        "signupAttributes": [
+          "EMAIL"
+        ],
+        "passwordProtectionSettings": {
+          "passwordPolicyMinLength": 8,
+          "passwordPolicyCharacters": []
+        },
+        "mfaConfiguration": "OFF",
+        "mfaTypes": [
+          "SMS"
+        ],
+        "verificationMechanisms": [
+          "EMAIL"
+        ]
+      },
+      "providerMetadata": {
+        "s3TemplateURL": "https://s3.amazonaws.com/amplify-myapp-dev-184113-deployment/amplify-cfn-templates/auth/myapp1099f97a-cloudformation-template.json",
+        "logicalId": "authmyapp1099f97a"
+      },
+      "lastPushTimeStamp": "2023-09-06T10:43:53.693Z",
+      "output": {
+        "UserPoolId": "ap-southeast-2_CoeSmUcXu",
+        "AppClientIDWeb": "4bd1stbm57cd94dfprtul0tlvf",
+        "AppClientID": "3q086shpjp7tr8kuhe4mm0jt5j",
+        "IdentityPoolId": "ap-southeast-2:e59c4e9a-00fd-46c7-ba14-6322231f61e2",
+        "UserPoolArn": "arn:aws:cognito-idp:ap-southeast-2:427176174668:userpool/ap-southeast-2_CoeSmUcXu",
+        "IdentityPoolName": "myapp1099f97a_identitypool_1099f97a__dev",
+        "UserPoolName": "myapp1099f97a_userpool_1099f97a"
+      },
+      "lastPushDirHash": "jJfqFOvjfF3wVj6YQCQyYnLkj6A="
+    }
+  }
+}
