+++ frontend/amplify/.config/local-aws-info.json
@@ -1,0 +1,7 @@
+{
+  "dev": {
+    "configLevel": "project",
+    "useProfile": false,
+    "awsConfigFilePath": "C:\\Users\\ngpin\\.amplify\\awscloudformation\\pOgkMmBTqF"
+  }
+}
