+++ frontend/amplify/.config/local-env-info.json
@@ -1,0 +1,5 @@
+{
+  "projectPath": "C:\\Users\\ngpin\\OneDrive - Singapore Management University\\Y3T1\\[5%] CS203 Collaborative Software Development\\[35%] Project\\testing\\my-app",
+  "defaultEditor": "Visual Studio Code",
+  "envName": "dev"
+}
