+++ frontend/amplify/backend/types/amplify-dependent-resources-ref.d.ts
@@ -1,0 +1,13 @@
+export type AmplifyDependentResourcesAttributes = {
+  "auth": {
+    "myapp1099f97a": {
+      "AppClientID": "string",
+      "AppClientIDWeb": "string",
+      "IdentityPoolId": "string",
+      "IdentityPoolName": "string",
+      "UserPoolArn": "string",
+      "UserPoolId": "string",
+      "UserPoolName": "string"
+    }
+  }
+}
