+++ frontend/amplify/team-provider-info.json
@@ -1,0 +1,20 @@
+{
+  "dev": {
+    "awscloudformation": {
+      "AuthRoleName": "amplify-myapp-dev-184113-authRole",
+      "UnauthRoleArn": "arn:aws:iam::427176174668:role/amplify-myapp-dev-184113-unauthRole",
+      "AuthRoleArn": "arn:aws:iam::427176174668:role/amplify-myapp-dev-184113-authRole",
+      "Region": "ap-southeast-2",
+      "DeploymentBucketName": "amplify-myapp-dev-184113-deployment",
+      "UnauthRoleName": "amplify-myapp-dev-184113-unauthRole",
+      "StackName": "amplify-myapp-dev-184113",
+      "StackId": "arn:aws:cloudformation:ap-southeast-2:427176174668:stack/amplify-myapp-dev-184113/e8d801a0-4ca1-11ee-b4d9-0abcacc0f598",
+      "AmplifyAppId": "d1pufa9ecg46gu"
+    },
+    "categories": {
+      "auth": {
+        "myapp1099f97a": {}
+      }
+    }
+  }
+}
