+++ frontend/src/Utils/constants.js
@@ -1,0 +1,2 @@
+/** Auth user token key */
+export const AUTH_USER_TOKEN_KEY = 'ReactAmplify.TokenKey';
